refactor, introduce a thread-local MockClock
